完成档案界面 完成银行卡界面（拖过来的） 新增SDWebImageView

+++ Lywsale/Lywsale/App/Personal（个人中心相关）/Footprint（足迹）/ViewController/FootprintViewController.m
@@ -30,7 +30,7 @@
     [self.view addSubview:footprintHeaderView];
     
     [footprintHeaderView setSelectIndex:^(NSInteger index) {
-        NSLog(@"%zd", index);
+        NSLog(@"%ld", (long)index);
     }];
 }
